requirements.txt
fortnitepy
websockets
pyyaml
mysql-connector-python
